refactor/ev-839/refactoring code
