+++ lib/modules/details/model/market_chart_view_data.dart
@@ -1,5 +1,5 @@
-class MarketChartViewData {
   // ignore: non_constant_identifier_names
+class MarketChartViewData {
   final List<List<num>> prices;
 
   MarketChartViewData({required this.prices});
