+++ lib/modules/details/widgets/details_line_chart.dart
@@ -17,9 +17,11 @@ class DetailsLineChart extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    int marketChartDay = ref.watch(chartDayProvider);
     List<List<num>> spots = marketChartData.prices;
-    double currentPrice = ref.watch(cryptoPriceProvider);
+    int marketChartDay = spots.length - 1 - ref.watch(chartDayProvider);
+    // double currentPrice = ref.watch(cryptoPriceProvider);
+    double currentPrice = spots[marketChartDay][1].toDouble();
+    // print(test);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
@@ -86,5 +88,3 @@ class DetailsLineChart extends HookConsumerWidget {
     );
   }
 }
-
-
