+++ lib/modules/wallet/model/crypto_data_view_data.dart
@@ -1,20 +1,18 @@
+// ignore_for_file: non_constant_identifier_names
+
 class CryptoDataViewData {
   final String id;
   final String symbol;
   final String name;
   final String image;
-  // ignore: non_constant_identifier_names
   final double current_price;
-  // ignore: non_constant_identifier_names
   final double market_cap_change_percentage_24h;
   CryptoDataViewData({
     required this.id,
     required this.symbol,
     required this.name,
     required this.image,
-    // ignore: non_constant_identifier_names
     required this.current_price,
-    // ignore: non_constant_identifier_names
     required this.market_cap_change_percentage_24h,
   });
