+++ lib/modules/wallet/model/crypto_list_view_data.dart
