+++ lib/modules/wallet/model/user_wallet_model.dart
