+++ lib/modules/wallet/views/wallet_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../../../shared/model/user_wallet_model.dart';
+import '../model/user_wallet_model.dart';
 import '../../../shared/providers/providers.dart';
+import '../providers/providers.dart';
 import '../widgets/crypto_list_view.dart';
 import '../widgets/header.dart';
 
@@ -14,7 +15,7 @@ class WalletScreen extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     List<UserWalletModel> userCryptoWallet = [];
     final userWallet = ref.watch(cryptoWalletProvider);
-    userWallet.whenOrNull(
+    return userWallet.when(
       data: (data) {
         userCryptoWallet = data.map((item) {
           return UserWalletModel(
@@ -22,15 +23,23 @@ class WalletScreen extends HookConsumerWidget {
             userCryptoBalance: item.userCryptoBalance,
           );
         }).toList();
+
+        return SafeArea(
+          child: Column(
+            children: <Widget>[
+              Header(userCryptoWallet: userCryptoWallet),
+              CryptoListView(userCryptoWallet: userCryptoWallet),
+            ],
+          ),
+        );
       },
-    );
-    return SafeArea(
-      child: Column(
-        children: <Widget>[
-          Header(userCryptoWallet: userCryptoWallet),
-          CryptoListView(userCryptoWallet: userCryptoWallet),
-        ],
-      ),
+      error: (error, stackTrace) =>
+          Text("The following error was found on UserWallet: $error"),
+      loading: () {
+        return const Center(
+          child: CircularProgressIndicator(),
+        );
+      },
     );
   }
 }
