+++ lib/shared/api/crypto_data/model/crypto_data_response.dart
@@ -1,28 +1,20 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'dart:convert';
 
-import 'package:json_annotation/json_annotation.dart';
-
-
-part 'crypto_data_response.g.dart';
-
-@JsonSerializable()
 class CryptoDataResponse {
   final String id;
   final String symbol;
   final String name;
   final String image;
-  // ignore: non_constant_identifier_names
   final double current_price;
-  // ignore: non_constant_identifier_names
   final double market_cap_change_percentage_24h;
   CryptoDataResponse({
     required this.id,
     required this.symbol,
     required this.name,
     required this.image,
-    // ignore: non_constant_identifier_names
     required this.current_price,
-    // ignore: non_constant_identifier_names
     required this.market_cap_change_percentage_24h,
   });
